Move `DocumentListItem` into `domain`
