+++ source/docq/domain.py
@@ -1,6 +1,9 @@
 """Domain classes for Docq."""
 
+import logging as log
+import sys
 from dataclasses import dataclass
+from datetime import datetime
 from typing import Any, Optional
 
 from .config import FeatureType, SpaceType
@@ -51,3 +54,41 @@ class ConfigKey:
     is_optional: bool = False
     is_secret: bool = False
     ref_link: str = None
+
+
+@dataclass
+class DocumentListItem:
+    """A tuple containing information about a document. These entries are used to create the document list for rendering UI."""
+
+    link: str
+    size: int
+    last_indexed: int
+
+    @staticmethod
+    def create_instance(document_link: str, document_text: str, indexed_on: int = None) -> "DocumentListItem":
+        """Creates a tuple containing information about a document.
+
+        Args:
+        document_link (str): The link to the document.
+        document_text (str): The text of the document.
+        indexed_on (Optional[int]): The timestamp of when the document was indexed. Defaults to `utcnow()`.
+
+        Returns:
+        DocumentListItem: A namedtuple containing the document link, indexed timestamp (epoch), and size in bytes.
+        """
+        try:
+            size_in_bytes = sys.getsizeof(document_text)
+
+            size_in_bytes = size_in_bytes if size_in_bytes > 0 else 0
+
+            if indexed_on is None:
+                indexed_on = datetime.timestamp(datetime.now().utcnow())
+
+            item = DocumentListItem(document_link, size_in_bytes, indexed_on)
+            log.debug("Created document list item: %s", item)
+            return item
+        except Exception as e:
+            log.error(
+                "Error creating document list item with %s %s %d",
+            )
+            raise e
